Update PubMedAPI.php
When a query string such as "allergy symptoms for demodex" is passed to the pubmed_esearch() function, only the first word from the query was being used to look up PubMed articles, i.e. "allergy". This was caused because the term was not being properly encoded. I added the urlencode() call to the term.